tests/Application.FunctionalTests/TodoItems/Commands/UpdateTodoItemTests.cs
#region

using TheHub.Application.TodoItems.Commands.CreateTodoItem;
using TheHub.Application.TodoItems.Commands.UpdateTodoItem;
using TheHub.Application.TodoLists.Commands.CreateTodoList;
using TheHub.Domain.Entities;

#endregion

namespace TheHub.Application.FunctionalTests.TodoItems.Commands;

#region

using static Testing;

#endregion

public class UpdateTodoItemTests : BaseTestFixture
{
    [Test]
    public async Task ShouldRequireValidTodoItemId()
    {
        UpdateTodoItemCommand command = new UpdateTodoItemCommand { Id = 99, Title = "New Title" };
        await FluentActions.Invoking(() => SendAsync(command)).Should().ThrowAsync<NotFoundException>();
    }

    [Test]
    public async Task ShouldUpdateTodoItem()
    {
        ulong userId = await RunAsDefaultUserAsync();

        int listId = await SendAsync(new CreateTodoListCommand { Title = "New List" });

        int itemId = await SendAsync(new CreateTodoItemCommand { ListId = listId, Title = "New Item" });

        UpdateTodoItemCommand command = new UpdateTodoItemCommand { Id = itemId, Title = "Updated Item Title" };

        await SendAsync(command);

        TodoItem? item = await FindAsync<TodoItem>(itemId);

        item.Should().NotBeNull();
        item!.Title.Should().Be(command.Title);
        item.LastModifiedBy.Should().NotBe(0);
        item.LastModifiedBy.Should().Be(userId);
        item.LastModified.Should().BeCloseTo(DateTime.Now, TimeSpan.FromMilliseconds(10000));
    }
}
